+++ examples/foundational/20d-persistent-context-gemini.py
@@ -237,7 +237,7 @@ async def main():
             voice_id="79a125e8-cd45-4c13-8a67-188112f4dd22",  # British Lady
         )
 
-        llm = GoogleLLMService(model="gemini-1.5-flash-latest", api_key=os.getenv("GOOGLE_API_KEY"))
+        llm = GoogleLLMService(model="gemini-2.0-flash-001", api_key=os.getenv("GOOGLE_API_KEY"))
 
         # you can either register a single function for all function calls, or specific functions
         # llm.register_function(None, fetch_weather_from_api)