+++ examples/foundational/25-google-audio-in.py
@@ -292,7 +292,7 @@ async def main():
 
         conversation_llm = GoogleLLMService(
             name="Conversation",
-            model="gemini-1.5-flash-latest",
+            model="gemini-2.0-flash-001",
             # model="gemini-exp-1121",
             api_key=os.getenv("GOOGLE_API_KEY"),
             # we can give the GoogleLLMService a system instruction to use directly
@@ -303,7 +303,7 @@ async def main():
 
         input_transcription_llm = GoogleLLMService(
             name="Transcription",
-            model="gemini-1.5-flash-latest",
+            model="gemini-2.0-flash-001",
             # model="gemini-exp-1121",
             api_key=os.getenv("GOOGLE_API_KEY"),
             system_instruction=transcriber_system_message,